Add dp_mode til SI SM
